Fixes to build with gcc4, mostly function prototypes and missing headers.
ok robert kettenis jsg

+++ gnu/usr.bin/groff/refer/refer.cc
@@ -513,7 +513,7 @@ static void do_file(const char *filename)
 	  if ((flags & FORCE_LEFT_BRACKET) || !have_text)
 	    pending_line += PRE_LABEL_MARKER;
 	  pending_line += pre;
-	  char lm = LABEL_MARKER + lt;
+	  char lm = LABEL_MARKER + (int)lt;
 	  pending_line += lm;
 	  pending_line += post;
 	  if ((flags & FORCE_RIGHT_BRACKET) || !have_text)
